[WiP] smbios_table.machine_type: filter machine types
The machine_type variant of smbios test attempts to run guests with
all machine types reported by `qemu -M ?`.  But, the code that
generates the QEMU command lines are not that smart (complex?) to
deal with those "sudden" changes.

The best solutions I can think of at this point is to add a filtering
mechanism, so that only machine types that match the given parameter
set by "traversal_machine_filter_param" (defaults to the standard
"machine_type" param) will be used.

In theory, an Avocado execution like:

 $ avocado run smbios_table.machine_type \
   --vt-extra-params traversal_machine_filter_param=machine_type \
                     machine_type=i440fx

Will filter out machine types such as q35 and avoid the command
line creation error.

Signed-off-by: Cleber Rosa <crosa@redhat.com>